accounts/forms.py
from django import forms
from django.contrib.auth import authenticate, login, get_user_model
from django.contrib.auth.forms import ReadOnlyPasswordHashField
from django.urls import reverse
from django.utils.safestring import mark_safe

User = get_user_model()

from .models import EmailActivation, Profile
from phonenumber_field.modelfields import PhoneNumberField
from localflavor.us.forms import USSocialSecurityNumberField, USStateField, USZipCodeField 
from localflavor.us.us_states import STATE_CHOICES
from django_countries.widgets import CountrySelectWidget


# write forms below
class ReactivateEmailForm(forms.Form):
    email       = forms.EmailField()

    def clean_email(self):
        email = self.cleaned_data.get('email')
        qs = EmailActivation.objects.email_exists(email) 
        if not qs.exists():
            register_link = reverse("register")
            msg = """This email does not exists, would you like to <a href="{link}">register</a>?
            """.format(link=register_link)
            raise forms.ValidationError(mark_safe(msg))
        return email

class UserAdminCreationForm(forms.ModelForm):
    """A form for creating new users. Includes all the required
    fields, plus a repeated password."""
    password1 = forms.CharField(label='Password', widget=forms.PasswordInput)
    password2 = forms.CharField(label='Password confirmation', widget=forms.PasswordInput)

    class Meta:
        model = User
        fields = ('first_name', 'middle_name', 'surname', 'email') #'full_name',)

    def clean_password2(self):
        # Check that the two password entries match
        password1 = self.cleaned_data.get("password1")
        password2 = self.cleaned_data.get("password2")
        if password1 and password2 and password1 != password2:
            raise forms.ValidationError("Passwords don't match")
        return password2

    def save(self, commit=True):
        # Save the provided password in hashed format
        user = super(UserAdminCreationForm, self).save(commit=False)
        user.set_password(self.cleaned_data["password1"])
        if commit:
            user.save()
        return user

class UserAdminChangeForm(forms.ModelForm):
    """A form for updating users. Includes all the fields on
    the user, but replaces the password field with admin's
    password hash display field.
    """
    password = ReadOnlyPasswordHashField() #can be chnaged to view passwords online in the admin

    class Meta:
        model = User
        fields = ('first_name', 'middle_name', 'surname', 'password', 'is_active', 'is_staff', 'is_client', 'is_superuser')

    def clean_password(self):
        # Regardless of what the user provides, return the initial value.
        # This is done here, rather than on the field, because the
        # field does not have access to the initial value
        return self.initial["password"]

class UserDetailChangeForm(forms.ModelForm):
    # company_name = forms.CharField(label='Name', required=False, widget=forms.TextInput(attrs={"class": 'form-control'}))

    class Meta:
        model = User
        fields = ('first_name', 'middle_name', 'surname')

class UserRegisterForm(forms.ModelForm):
    """A form for creating new users. Includes all the required
    fields, plus a repeated password."""
    password1 = forms.CharField(label='Password', widget=forms.PasswordInput)
    password2 = forms.CharField(label='Password confirmation', widget=forms.PasswordInput)

    class Meta:
        model = User
        fields = ('first_name', 'middle_name', 'surname', 'email') #'full_name',)

    def clean_password2(self):
        # Check that the two password entries match
        password1 = self.cleaned_data.get("password1")
        password2 = self.cleaned_data.get("password2")
        if password1 and password2 and password1 != password2:
            raise forms.ValidationError("Passwords don't match")
        return password2

    def save(self, commit=True):
        # Save the provided password in hashed format
        user = super(UserRegisterForm, self).save(commit=False)
        user.set_password(self.cleaned_data["password1"])
        user.is_active = False # send confirmation email via signals
        user.is_client = True
        # obj = EmailActivation.objects.create(user=user)
        # obj.send_activation_email()
        if commit:
            user.save()
        return user

class ProfileEditForm(forms.ModelForm):
    SEX = (
        ('s', 'SEX'),
        ('m', 'MALE'),
        ('f', 'FEMALE'),
    )
    MARITAL = (
        ('si', 'SINGLE'),
        ('ma', 'MARRIED'),
        ('di', 'DIVORCED'),
        ('se', 'SEPERATED'),
    )
    KIDS = (
        ('no', 'NO'),
        ('yes', 'YES'),
        ('maybe', 'MAYBE')
    )
    SEX_INTEREST = (
        ('m', 'MALE'),
        ('f', 'FEMALE'),
        ('b', 'MALE & FEMALE'),
    )
    PROFILE_STATE_CHOICES = list(STATE_CHOICES)
    PROFILE_STATE_CHOICES.insert(0, ('', 'STATES'))
    sex = forms.CharField(widget=forms.Select(choices=SEX))
    marital = forms.CharField(widget=forms.Select(choices=MARITAL))
    kids = forms.CharField(widget=forms.Select(choices=KIDS))
    interest = forms.CharField(widget=forms.Select(choices=SEX_INTEREST))
    DOB = forms.DateField(widget=forms.TextInput(attrs={
        'class': 'datepicker'
    }))
    zip_code = USZipCodeField()
    state_province = USStateField(widget=forms.Select(choices=PROFILE_STATE_CHOICES))
    SSN = USSocialSecurityNumberField()
    phone = PhoneNumberField()
    fax = PhoneNumberField()

    class Meta:
        model = Profile
        fields = ('username', 'passport', 'sex', 'interest', 'marital', 'kids', 'next_of_kin', 'height', 'DOB', 'phone', 'fax', 'address', 'city', 'zip_code', 'state_province', 'SSN', 'country')
        widgets = {'country': CountrySelectWidget()}

class LoginForm(forms.Form):
    email    = forms.EmailField(label='Email')
    password = forms.CharField(widget=forms.PasswordInput)

    def __init__(self, request, *args, **kwargs):
        self.request = request
        super(LoginForm, self).__init__(*args, **kwargs)

    def clean(self):
        request = self.request
        data = self.cleaned_data
        email  = data.get("email")
        password  = data.get("password")
        qs = User.objects.filter(email=email)
        if qs.exists():
            # user email is registered, check active/
            not_active = qs.filter(is_active=False)
            if not_active.exists():
                ## not active, check email activation
                link = reverse("account:resend-activation")
                reconfirm_msg = """Go to <a href='{resend_link}'>
                resend confirmation email</a>.
                """.format(resend_link = link)
                confirm_email = EmailActivation.objects.filter(email=email)
                is_confirmable = confirm_email.confirmable().exists()
                if is_confirmable:
                    msg1 = "Please check your email to confirm your account or " + reconfirm_msg.lower()
                    raise forms.ValidationError(mark_safe(msg1))
                email_confirm_exists = EmailActivation.objects.email_exists(email).exists()
                if email_confirm_exists:
                    msg2 = "Email not confirmed. " + reconfirm_msg
                    raise forms.ValidationError(mark_safe(msg2))
                if not is_confirmable and not email_confirm_exists:
                    raise forms.ValidationError("This user is inactive.")
        user = authenticate(request, username=email, password=password)
        if user is None:
            raise forms.ValidationError("Invalid credentials")
        login(request, user)
        self.user = user
        return data

    # def form_valid(self, form):
    #     request = self.request
    #     next_ = request.GET.get('next')
    #     next_post = request.POST.get('next')
    #     redirect_path = next_ or next_post or None
    #     email  = form.cleaned_data.get("email")
    #     password  = form.cleaned_data.get("password")
        
    #     print(user)
    #     if user is not None:
    #         if not user.is_active:
    #             print('inactive user..')
    #             messages.success(request, "This user is inactive")
    #             return super(LoginView, self).form_invalid(form)
    #         login(request, user)
    #         user_logged_in.send(user.__class__, instance=user, request=request)
    #         try:
    #             del request.session['guest_email_id']
    #         except:
    #             pass
    #         if is_safe_url(redirect_path, request.get_host()):
    #             return redirect(redirect_path)
    #         else:
    #             return redirect("/")
    #     return super(LoginView, self).form_invalid(form)

