requirements.txt
Babel==2.6.0
boto==2.49.0
boto3==1.9.95
botocore==1.12.95
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-localflavor==2.1
django-phonenumber-field==2.2.0
django-storages==1.7.1
docutils==0.14
idna==2.8
jmespath==0.9.3
phonenumbers==8.10.5
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
